rewrote teleop and auto to use servo turret
TODO: rewrite auto tracking to use servo turret

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mechanisms/shooter/Turret.java
@@ -177,7 +177,7 @@ public class Turret extends HwMotor {
         resetController.setTargetPosition(0);
         setDirection(DcMotorSimple.Direction.FORWARD);
         if (Globals.isTeleOp && Globals.turretStartPos > 1)
-            setEncoderBase(getEncoder().getPosition() - Globals.turretStartPos);
+            setEncoderBase((int) (getEncoder().getPosition() - Globals.turretStartPos));
         else
             setEncoderBase(getEncoder().getPosition());
         Globals.turretStartPos = 0;
